+++ src/solve.jl
@@ -1282,32 +1282,24 @@ function promote_f(f::F, ::Val{specialize}, u0, p, t) where {F, specialize}
         f = @set f.jac_prototype = similar(f.jac_prototype, uElType)
     end
 
-    @static if VERSION >= v"1.8-"
-        f = if f isa ODEFunction && isinplace(f) && !(f.f isa AbstractSciMLOperator) &&
-               # Some reinitialization code still uses NLSolvers stuff which doesn't
-               # properly tag, so opt-out if potentially a mass matrix DAE
-               f.mass_matrix isa UniformScaling &&
-               # Jacobians don't wrap, so just ignore those cases
-               f.jac === nothing &&
-               ((specialize === SciMLBase.AutoSpecialize && eltype(u0) !== Any &&
-                 RecursiveArrayTools.recursive_unitless_eltype(u0) === eltype(u0) &&
-                 one(t) === oneunit(t) &&
-                 hasmethod(ArrayInterface.promote_eltype,
-                     Tuple{Type{typeof(u0)}, Type{dualgen(eltype(u0))}}) &&
-                 hasmethod(promote_rule,
-                     Tuple{Type{eltype(u0)}, Type{dualgen(eltype(u0))}}) &&
-                 hasmethod(promote_rule,
-                     Tuple{Type{eltype(u0)}, Type{typeof(t)}})) ||
-                (specialize === SciMLBase.FunctionWrapperSpecialize &&
-                 !(f.f isa FunctionWrappersWrappers.FunctionWrappersWrapper)))
-            return unwrapped_f(f, wrapfun_iip(f.f, (u0, u0, p, t)))
-        else
-            return f
-        end
+    f = if f isa ODEFunction && isinplace(f) && !(f.f isa AbstractSciMLOperator) &&
+            # Some reinitialization code still uses NLSolvers stuff which doesn't
+            # properly tag, so opt-out if potentially a mass matrix DAE
+            f.mass_matrix isa UniformScaling &&
+            # Jacobians don't wrap, so just ignore those cases
+            f.jac === nothing &&
+            ((specialize === SciMLBase.AutoSpecialize && eltype(u0) !== Any &&
+                RecursiveArrayTools.recursive_unitless_eltype(u0) === eltype(u0) &&
+                one(t) === oneunit(t) && hasdualpromote(u0,t)) ||
+            (specialize === SciMLBase.FunctionWrapperSpecialize &&
+                !(f.f isa FunctionWrappersWrappers.FunctionWrappersWrapper)))
+        return unwrapped_f(f, wrapfun_iip(f.f, (u0, u0, p, t)))
     else
         return f
     end
 end
+
+hasdualpromote(u0,t) = true
 
 function promote_f(f::SplitFunction, ::Val{specialize}, u0, p, t) where {specialize}
     typeof(f.cache) === typeof(u0) && isinplace(f) ? f : remake(f, cache = zero(u0))
@@ -1462,7 +1454,7 @@ function check_prob_alg_pairing(prob, alg)
         throw(ProblemSolverPairingError(prob, alg))
     end
 
-    if isdefined(prob, :u0) && eltype(prob.u0) <: ForwardDiff.Dual &&
+    if isdefined(prob, :u0) && eltypedual(prob.u0) &&
        !SciMLBase.isautodifferentiable(alg)
         throw(DirectAutodiffError())
     end
